+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/lrop_v4/webapp/test/integration/FirstJourney.js
@@ -9,9 +9,7 @@ sap.ui.define([
 
         opaTest("Start application", function (Given, When, Then) {
             Given.iStartMyApp();
-
             Then.onTheTravelList.iSeeThisPage();
-
         });