+++ packages/generator-odata-downloader/CHANGELOG.md
@@ -1,4 +1,22 @@
-# @sap-ux/generator-odata-download
+# @sap-ux/generator-odata-downloader
+
+## 0.0.10
+
+### Patch Changes
+
+- c53a4ba: chore(generator-odata-downloader): upgrade shared devDependencies (jest 30, i18next 25, @types/yeoman-generator 5.2.14)
+
+## 0.0.9
+
+### Patch Changes
+
+- 55d833f: fix i18next init showSupportNotice: false.
+
+## 0.0.8
+
+### Patch Changes
+
+- 42b558b: Updates readme. Move deps tto devDeps
 
 ## 0.0.7
 
@@ -54,4 +72,4 @@
 
 ### Patch Changes
 
-- b7461bc: Adds new package @sap-ux/generator-odata-download
+- b7461bc: Adds new package @sap-ux/generator-odata-downloader